feat: rename boardConfig to gameConfig
