+++ src/components/__tests__/GoBoard.test.tsx
@@ -5,7 +5,7 @@ import {
   BOARD_CONFIG,
   BOARD_DIMENSIONS,
   StoneColor,
-} from '@/constants/boardConfig';
+} from '@/constants/gameConfig';
 import { useMove } from '@/hooks/move';
 import type { Position, Move } from '@/hooks/move';
